+++ ChangeLog.rst
@@ -8,7 +8,7 @@ Unreleased
 Added
 +++++
 
-* Add functions relating to exponentiation: exp, log, logn, log10, power
+* Add functions relating to exponentiation: exp, expb, exp10, log, logb, log10, power
 * Add core comparison functions for equality, closeness, finiteness
 
 v0.3 - 2018-03-31