+++ ChangeLog.rst
@@ -5,13 +5,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Unreleased
 ----------
 
+v0.4.1 - 2018-04-08
+-----------------
+
+Fixed
++++++
+
+* Exponential for bases other than e are calculated correctly.
+
+v0.4.0 - 2018-04-08
+-----------------
+
 Added
 +++++
 
 * Add functions relating to exponentiation: exp, expb, exp10, log, logb, log10, power
 * Add core comparison functions for equality, closeness, finiteness
 
-v0.3 - 2018-03-31
+v0.3.0 - 2018-03-31
 -----------------
 
 Added
@@ -27,7 +38,7 @@ Changed
 * Converting from Euler now takes alpha, beta, and gamma as separate args
 * Ensure more complete coverage
 
-v0.2 - 2018-03-08
+v0.2.0 - 2018-03-08
 -----------------
 
 Added
@@ -44,7 +55,7 @@ Changed
 * Clean up CI.
 * Ensure pep8 compliance.
 
-v0.1 - 2018-02-26
+v0.1.0 - 2018-02-26
 -----------------
 
 Added